services/bu5_care_operations/pyproject.toml
[project]
name = "ritecare-bu5-care-operations"
version = "0.1.0"
description = "RiteCare BU5 - Care Operations microservice"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "motor>=3.5",
    "pymongo>=4.8",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "slowapi>=0.1.9",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "aiokafka>=0.11",
    "slack-sdk>=3.30",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "google-genai>=1.0",
    
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "service", "dao", "common"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
